chat/requirements.txt
#openai~=1.65.4
#openai>=1.0.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
streamlit>=1.36.0

